# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MartHub API",
    "version": "1.0.0",
    "description": "Implemented foundation, authentication, account, and public catalog endpoints. Future inventory: docs/API.md."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
    "/auth/logout": {
      "post": {
        "operationId": "authLogout",
        "description": "Idempotently revoke the cookie family, including rotated successors.",
        "security": [
          {},
          {
            "RefreshCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Origin"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyInput"
              },
              "example": {}
            }
          }
        },
        "responses": {
          "204": {
            "$ref": "#/components/responses/AuthLoggedOut"
          },
          "400": {
            "$ref": "#/components/responses/AuthError"
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "413": {
            "$ref": "#/components/responses/AuthError"
          },
          "415": {
            "$ref": "#/components/responses/AuthError"
          },
          "422": {
            "$ref": "#/components/responses/AuthError"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      }
    },
    "/auth/logout-all": {
      "post": {
        "operationId": "authLogoutAll",
        "description": "Revoke all families for the authenticated user.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Origin"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyInput"
              },
              "example": {}
            }
          }
        },
        "responses": {
          "204": {
            "$ref": "#/components/responses/AuthLoggedOut"
          },
          "400": {
            "$ref": "#/components/responses/AuthError"
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "413": {
            "$ref": "#/components/responses/AuthError"
          },
          "415": {
            "$ref": "#/components/responses/AuthError"
          },
          "422": {
            "$ref": "#/components/responses/AuthError"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      }
    },
...
    "/auth/refresh": {
      "post": {
        "operationId": "authRefresh",
        "description": "No body or empty JSON object. 300 attempts/IP/15 minutes. Atomic rotation; replay revokes the family including concurrent successors.",
        "security": [
          {
            "RefreshCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/Origin"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyInput"
              },
              "example": {}
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/AuthSession"
          },
          "400": {
            "$ref": "#/components/responses/AuthError"
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "413": {
            "$ref": "#/components/responses/AuthError"
          },
          "415": {
            "$ref": "#/components/responses/AuthError"
          },
          "422": {
            "$ref": "#/components/responses/AuthError"
          },
          "429": {
            "$ref": "#/components/responses/AuthRateLimited"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message", "requestId"],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string",
                "pattern": "^[A-Za-z0-9_-]{1,64}$"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      },
...
      "AuthResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "accessToken": {
                "type": "string",
                "minLength": 1
              },
              "expiresIn": {
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              },
              "user": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": ["CUSTOMER", "ADMIN"]
                  },
                  "phone": {
                    "type": ["string", "null"]
                  },
                  "status": {
                    "type": "string",
                    "enum": ["ACTIVE", "SUSPENDED", "ARCHIVED"]
                  }
                },
                "required": ["id", "email", "firstName", "lastName", "role"],
                "additionalProperties": false
              }
            },
            "required": ["accessToken", "expiresIn", "user"],
            "additionalProperties": false
          }
        },
        "required": ["data"],
        "additionalProperties": false
      },
      "EmptyInput": {
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "parameters": {
      "RequestId": {
        "in": "header",
        "name": "X-Request-Id",
        "schema": {
          "type": "string",
          "pattern": "^[A-Za-z0-9_-]{1,64}$"
        },
        "description": "Invalid values are replaced with a generated UUID."
      },
      "Origin": {
        "in": "header",
        "name": "Origin",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uri"
        },
        "description": "Must exactly equal WEB_ORIGIN. HTTPS required in production.",
        "example": "https://shop.example.test"
      },
...
    },
    "responses": {
      "AuthLoggedOut": {
        "description": "Revocation committed; cookie cleared. No response body.",
        "headers": {
          "Cache-Control": {
            "schema": {
              "const": "no-store"
            }
          },
          "X-Request-Id": {
            "schema": {
              "type": "string"
            }
          },
          "Set-Cookie": {
            "schema": {
              "type": "string"
            },
            "example": "mh_refresh=; Path=/api/v1/auth; Expires=Thu, 01 Jan 1970 00:00:00 GMT; HttpOnly; Secure; SameSite=Lax"
          }
        }
      },
...
      "AuthError": {
        "description": "Normalized error: INVALID_CREDENTIALS/UNAUTHORIZED (401), ORIGIN_NOT_ALLOWED (403), EMAIL_IN_USE (409), VALIDATION_ERROR (422).",
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "no-store"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "AuthRateLimited": {
        "description": "RATE_LIMITED: shared PostgreSQL attempt budget exceeded.",
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "no-store"
            }
          },
          "Retry-After": {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 900
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "AuthSession": {
        "description": "Access token lifetime is 900 seconds. Keep access tokens in browser memory. No refresh token or password hash in JSON.",
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "no-store"
            }
          },
          "Set-Cookie": {
            "description": "Host-only mh_refresh; HttpOnly; SameSite=Lax; Path=/api/v1/auth; Secure in production; 30-day expiry. Refresh 401 clears this cookie.",
            "schema": {
              "type": "string"
            },
            "example": "mh_refresh=REDACTED; Path=/api/v1/auth; Expires=Wed, 14 Oct 2026 00:00:00 GMT; HttpOnly; Secure; SameSite=Lax"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AuthResponse"
            },
            "example": {
              "data": {
                "accessToken": "REDACTED_ACCESS_TOKEN",
                "expiresIn": 900,
                "user": {
                  "id": "b2ecb79a-b74a-4748-93dc-f2fc02b93b3d",
                  "email": "customer@example.test",
                  "firstName": "Minh",
                  "lastName": "Nguyen",
                  "role": "CUSTOMER"
                }
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Active session and current database user status/role required."
      },
      "RefreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mh_refresh",
        "description": "Rotated opaque token. Only SHA-256 hashes are persisted; replay revokes the family."
      }
    }
  }
}